move time range change subscriptions to timescale API.
